requirements.txt
pandas>=2.2,<3.0
plotly>=5.24,<6.0
pyyaml>=6.0,<7.0
pytest>=8.3,<9.0
